// src/enums/service.enum.ts
export enum ServicesEnum {
    RESIDENTIAL_CLEANING_SERVICES = "Residential Cleaning Services",
    COMMERCIAL_CLEANING_SERVICES = "Commercial Cleaning Services",
    MOVE_IN_MOVE_OUT_CLEANING = "Move-In/Move-Out Cleaning",
    POST_CONSTRUCTION_CLEANING = "Post-Construction Cleaning",
    EVENT_CLEANUP_SERVICES = "Event Cleanup Services",
    CLEANING_ORGANIZATION_CONSULTING = "Cleaning Organization Consulting",
    CARPET_CLEANING = "Carpet Cleaning",
    UPHOLSTERY = "Upholstery",
}